[BE] Fix:  캐릭터 해금 버그 수정

+++ backend/src/main/java/com/ssafy/gaese/domain/algorithm/application/AlgoService.java
@@ -408,7 +408,7 @@ public class AlgoService {
     }
 
 
-    void userCharacterSet(User user, int charId, Map<Long,Characters> characters)
+    void userCharacterSet(User user, long charId, Map<Long,Characters> characters)
     {
         UserCharacter userCharacter = new UserCharacter();
         userCharacter.setUser(user);
