+++ backend/src/main/java/com/ssafy/gaese/domain/cs/application/CsService.java
@@ -443,7 +443,7 @@ public class CsService {
     }
 
 
-    void userCharacterSet(User user, int charId, Map<Long,Characters> characters)
+    void userCharacterSet(User user, long charId, Map<Long,Characters> characters)
     {
         UserCharacter userCharacter = new UserCharacter();
         userCharacter.setUser(user);
