+++ backend/src/main/java/com/ssafy/gaese/domain/typing2/application/TypingService.java
@@ -310,7 +310,7 @@ public class TypingService {
 
     }
 
-    void userCharacterSet(User user, int charId, Map<Long,Characters> characters)
+    void userCharacterSet(User user, long charId, Map<Long,Characters> characters)
     {
         UserCharacter userCharacter = new UserCharacter();
         userCharacter.setUser(user);
