+++ backend/src/main/java/com/ssafy/gaese/domain/user/application/AttendanceService.java
@@ -121,7 +121,7 @@ public class AttendanceService {
     }
 
 
-    void userCharacterSet(User user, int charId, Map<Long,Characters> characters)
+    void userCharacterSet(User user, long charId, Map<Long,Characters> characters)
     {
         UserCharacter userCharacter = new UserCharacter();
         userCharacter.setUser(user);
